# docs/swagger.yaml
basePath: /
definitions:
...
  reqresp.ErrorInfo:
    properties:
      code:
        example: 400
        type: integer
        x-order: "1"
      details:
        example: 'Invalid ID format: expected integer, got ''abc'''
        type: string
        x-order: "2"
    type: object
...
  reqresp.StandardResponse:
    properties:
      data:
        x-order: "3"
      error:
        allOf:
        - $ref: '#/definitions/reqresp.ErrorInfo'
        x-order: "4"
      message:
        example: Operation completed successfully
        type: string
        x-order: "2"
      success:
        example: true
        type: boolean
        x-order: "1"
    type: object
...
info:
  contact: {}
  description: API documentation for Marketplace project
  title: Go Marketplace API
  version: "1.0"
paths:
...
  /api/products:
    get:
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reqresp.StandardResponse'
      summary: List all products
      tags:
      - products
  /api/products/{id}:
    get:
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reqresp.StandardResponse'
      summary: Get product with offers
      tags:
      - products
  /api/products/search:
    get:
      description: Search for products using full-text search. Returns 404 if no products
        found.
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reqresp.StandardResponse'
        "404":
          description: No products found
          schema:
            $ref: '#/definitions/reqresp.StandardResponse'
      summary: Search products
      tags:
      - products
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"